Versão 2.0.1

+++ src/main/java/br/com/smartmed/consultas/model/ConsultaModel.java
@@ -8,6 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -31,7 +32,7 @@ public class ConsultaModel {
 
     @Column(name = "valor",nullable = false)
     @NotNull(message = "O valor não poder ser nulo")
-    private float valor;
+    private BigDecimal valor;
 
     @Column(name = "observacoes",length = 1024,nullable = true)
     private String observacoes;
